+++ src/main.rs
@@ -1,2 +1,3 @@
+
 fn main() {
 }
